+++ CHANGES.rst
@@ -16,6 +16,10 @@
 - Fix bug where a dictionary containing a key ``id`` caused
   any contained references to fail to resolve [#1716]
 
+- Issue a ``AsdfManifestURIMismatchWarning`` during write if a used
+  extension was created from a manifest registered with a uri that
+  does not match the id in the manifest [#1785]
+
 
 3.2.0 (2024-04-05)
 ------------------